Exploit value tagging for faster int arithmetic.

+++ value.h
@@ -3,8 +3,8 @@
 
 #include "machine.h"
 
-typedef uint32_t Value;
-typedef int32_t  Int;
+typedef int32_t Value;
+typedef int32_t Int;
 
 // Stupid C with separate namespaces for structs and types
 typedef struct Obj         Obj;
@@ -35,7 +35,7 @@ typedef struct ObjUpvalue  ObjUpvalue;
 #define IS_NIL(value)    ((value) == NIL_VAL)
 #define IS_EMPTY(value)  ((value) == EMPTY_VAL)
 #define IS_INT(value)    ((value) & 1)
-#define IS_OBJ(value)    (((value) & 1) == 0 && ((value) > TRUE_VAL))
+#define IS_OBJ(value)    (((value) & 1) == 0 && ((unsigned)(value) > TRUE_VAL))
 #define IS_FALSEY(value) (((value) | 2) == FALSE_VAL)
 
 #define AS_BOOL(value)   ((value) == TRUE_VAL)
